created dataAccess.js module
