+++ scripts/DiamondSizes.js
@@ -1,4 +1,4 @@
-import { getSizes, setSize, getOrderBuilder} from "./database.js"
+import { getSizes, setSize, getOrderBuilder} from "./dataAccess.js"
 import { renderAllHTML } from "./main.js"
