+++ scripts/KneelDiamonds.js
@@ -4,7 +4,7 @@ import { JewelryStyles } from "./JewelryStyles.js"
 import { Orders } from "./Orders.js"
 import { Metals } from "./Metals.js"
 import { JewelryTypes } from "./JewelryTypes.js"
-import { addCustomOrder } from "./database.js"
+import { addCustomOrder } from "./dataAccess.js"
 
 
 document.addEventListener(
