+++ scripts/Metals.js
@@ -1,4 +1,4 @@
-import { getMetals, setMetal, getOrderBuilder } from "./database.js"
+import { getMetals, setMetal, getOrderBuilder } from "./dataAccess.js"
 import { renderAllHTML } from "./main.js"
